requirements.txt
ollama>=0.2.1
click>=8.1.7
rich>=13.7.0
pydantic>=2.5.0
pyyaml>=6.0.1
jinja2>=3.1.2
pathlib2>=2.3.7
tree-sitter>=0.20.4
tree-sitter-python>=0.20.4
tree-sitter-javascript>=0.20.3
tree-sitter-go>=0.20.0
asyncio-mqtt>=0.16.1
aiofiles>=23.2.1
typer>=0.9.0
colorama>=0.4.6
